Allow for default umodes to change at runtime
When default umodes are changed in ircd.conf, these changes will now be
applied to all local clients upon receiving a /rehash. We track which
umodes a user has explicitly (un)set and avoid modifying those umodes
when changing the defaults -- only umodes that a user didn't explicitly
touch are subject to change on rehash.