Revert "Fix WHOIS reply not respecting a lack of the multi-prefix capability"
A server handling a remote WHOIS cannot safely use IsCapable() to determine
if the client has multi-prefix enabled or not.

This is because client capabilities are not shared between servers; and the
IsCapable() macro dereferences source_p->localClient which will be NULL if
source_p is not on this server.

Instead rework the commit to always pretend that multi-prefix is not
enabled, to still respect the fact that multiple prefixes should not be
returned in case a client does not have multi-prefix enabled.

This reverts commit 4633665ad62bdfe26529