Don't set the oper flags on client
The oper_conf flags are a different set of flag
bits than the client flags. The oper flags are
hardcoded to have ENCRYPTED set which is the same
flag as PINGSENT on the client struct. This can
cause a client to timeout if the last thing it
does is oper up.