Add "bypass" capability to spamfilter
A new flag ACT_BYPASS was added as a generic flag to bypass +u if set.
Opers remain immune, both as senders and recipients, so that spamfilter
cannot be abused to lock people out of managing filters, talking to
services, etc.

Three new configuration items were added to the general section of
ircd.conf to control filter behaviour:

- filter_sees_user_info: This replaces the old #defines for FILTER_NICK,
  FILTER_IDENT, and FILTER_HOST to determine whether the hostmask of the
  user is passed to the spamfilter or whether we pass a dummy *!*@*.
- filter_bypass_all: If set, the new ACT_BYPASS can allow every filter
  action (DROP, KILL, ALARM) to bypass +u if set in conjunction with
  those flags. If unset, only the DROP action can bypass +u and
  ACT_BYPASS has no effect on other actions.
- filter_exit_message: This is the quit reason displayed, to move it out
  of a #define (we still keep a define around for a default reason in
  case it is unspecified).

If either filter_sees_user_info or filter_bypass_all are set, the
VERSION command will display a new 'F' flag to indicate the ability for
filter to spy on users.