Cleanup areas impacted by client tags
With the new TAGMSG support in m_message, filter now receives TAGMSG but
would choke on processing them due to not having a mapping from that
message type to a command name. Right now, no filtering is done on tag
contents because there is currently no tags that we support that take
arbitrary user-defined data (and thus would require filtering). This
fact does not change in the course of this patchset.

The client_tags code to register client tags for the CLIENTTAGSDENY
ISUPPORT message had some general formatting cleanup as well as a fix
for a segfault when removing support for a client tag if that client tag
was the last one being removed (causing the src and dest for strcpy to
be the same pointer) as well as a bugfix for the strcmp check that
would've caused it to overwrite the wrong client tag entry.