ircd/send.c: Better support for outbound tags
There are now sendto_* functions that take explicit tags to pass to the
outgoing MsgBuf, in addition to any hook functions that listen to
outbound_msgbuf to add tags of their own. The outbound_msgbuf hook
function now receives a fully-parsed line, meaning it will have a
defined parameters array and (sometimes) origin. Hook functions can make
use of this to read message context when determining whether to add
tags, and can also manipulate the outbound MsgBuf as it will be unparsed
back into an IRC line after hooks have executed to be sent off.
Previously, only the tags portion was used and outbound_msgbuf had no
capability to examine or manipulate the rest of the outgoing message.

These functions also make use of the new MsgBuf caching layer that takes
a local and remote source. For functions that receive a static source as
part of the format pattern, NULL is used so that the cache inherits the
source from the parsed line.

In some cases, a sendto function may produce completely different
messages depending on the situation (e.g. sendto_channel_opmod which
behaves differently depending on if the remote server supports EOPMOD or
not, and if not, whether or not the target channel is +m). In such
cases, separate MsgBuf caches with separate MsgBufs were used. The
outbound_msgbuf hook is only executed once however, and all tags are
copied over to the other MsgBufs. This allows things like msgid to be
stable regardless of which variation is being sent out to which client.

Internal static functions beginning with underscores were renamed to be
sendto_*_internal instead. Identifiers with leading underscores in file
scope are reserved by the C standard, and their usage here was undefined
behavior.