Update tests
A number of tests needed to be updated to account for the fact that tags
are allowed to be up to 8191 bytes (or 4094 bytes for the
client-specified portion or server-specified portion individually). The
numbers in these tests were updated, and in some cases we now test both
server and client tags to ensure consistent behavior between both types.