Use -c log.showsignature=false for git log
As explained in libera #solanum:

<h|weechat> Just discovered that Makefile.am:26 " datecode=`git log -1
--pretty=format:%ct`; \" (which gets fed into serno.h) silently
malfunctions and includes git signature-related data in the generated
serno.h in some conditions, including when the latest commit is signed,
if the git config option log.showsignature is set to true, breaking the
build process of solanum. Is a PR that fixes this welcome?

This change should not cause any regressions on old versions of git, as,
when tested on git 2.44.1, git appears to ignore invalid config
options.